test(gateway): split terminal.upload cases out of terminal method tests
terminal.test.ts sits over the 1000-line limit once upload cases grow, so the
upload handler gets its own file and the request-context builders move to
terminal.test-helpers.ts.
